pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.115.0",
    "google-genai>=1.59.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.1.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.4.0",
    "pymupdf>=1.24.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.30",
    "uvicorn[standard]>=0.30.0",
]
